Image Watermarks
Adds watermarks to graph images.

In the  lower left corner,  a hash of  the current
system configuration is  overlaid onto the images.
This ensures  that images resulting from  the same
systems  configuration can  be attributed  to each
other.

In  the lower  right corner,  a timestamp  for the
current program run is  overlaid. This is the same
timestamp which  is used  in the naming  scheme of
the graph images themselves. It allows to date the
images  even  if  they  are  renamed  as  well  as
facilitates dating them when  image file names are
not readily viewable (such  as when embedding them
on a website) for convenience.

NOTE: Currently, an error will  be produced by the
pnd module  of GD related to  interlacing. This is
not fatal and  as far as I can tell  at the moment
is a bug. Disregard without worry.

 On branch master

 Changes to be committed:
	new file:   Ltt/Digester.pm
	new file:   Ltt/Imaging.pm
	modified:   Ltt/Plotting.pm
	modified:   Ltt/Strings.pm
	modified:   ltt-rankings.pl
	modified:   ltt-rankings.txt